LuaRefactor: Fix typos in registerRegion

+++ src/map/lua/lua_zone.h
@@ -38,7 +38,7 @@ public:
         return m_pLuaZone;
     }
 
-    void        registerRegion(uint16 RegionID, float x1, float y1, float z1, float x2, float y2, float z2);
+    void        registerRegion(uint32 RegionID, float x1, float y1, float z1, float x2, float y2, float z2);
     sol::object levelRestriction();
     auto        getPlayers() -> std::vector<CLuaBaseEntity>;
     ZONEID      getID();
